progress: finish first set of refactoring
